Finalizing product
Created three types of experiments with slight changes. Added
instruction img to intro to the experiment for each experiment type.
Divs containing the explanation for which is the positive and which is
the negative color are now inline. Changed the number of experiments of
each percentage type to 30 instead of 1 that was put for testing.
Created folders for each instruction.